Fix store to have more consistent this binding
The store’s methods mostly use scoped variables like `values` and `callbacks`, but two lines referred to `this`. That is not immediately a problem, but it might become one later, when users do semi-weird stuff with wrapping the library and overriding store methods. The problem would be, that the scoped variable access is bound lexically while `this` is famously dynamic.
Inconsistent access to values in that sense might potentially lead to extremely hard to detect bugs for advanced users of the library.