Fix existing values being mistakenly overwritten
When no `initialValue` was provided, `create` would assign `undefined` to the key to be created even if that key already existed.

Additionally, this commit adds an exception that is thrown, when an `initialValue` is given for a key that already exists.
While it is more concise to create and override a value with `store.create('foo', 'bar')` instead of first creating and then writing to it with `store.create('foo'); store.foo = 'bar'`, I do think that the first is indistinguishible from an error in cases in which `foo` already has a value. To prevent this behavior, I decided to throw an exception here.