Fix subscribing to improperly created keys failing
Calling `subscribe` on a key that exists as a property on the store, but was not properly created with `create` causes an internal call of `create`. Since a recent addition of a check against potentially unwanted cases of existing properties getting overridden, this would fail. Fixed the issue by only throwing an error if the values are not strict equal, which they are, if `create` gets called from `subscribe`.