Add default store
Fleux now comes with a default store that stateful components use if the context does not contain a reference to a store. This way, components can use a state if no call to `connect` was made. This is generally not advised to be done and currently there is no way to get access to the default store.