Add store injection for empty `withState` call
It is now possible to call `withState` with no arguments and have the resulting factory return a component that gets the store reference injected into its props.

The store reference is retrieved via context just as before.

The name of the store reference is 'store'.

The `withState` call without arguments uses a completely different code path than the previously available call signatures.