Refactor getters of implicitly bound props
This improves the implementation of implicit prop binding by making it more resilient (and possibly also more performant).

The getter logic is moved to a special method on the store. This method is only accessible via a library-internal symbol.

I chose to use a store method instead of another `get`-trap because I wanted to pass the currently rendering component as an argument. The alternative was (before this commit), to write a reference to the currently rendering component to a central place for each component render and then remove the reference after rendering. This felt somewhat brittle as it relied heavily on the correct execution order of modules that have no real need to be synchronized.