Refactor implementation of explicitly bound props
This commit changes how explicitly bound props are bound to a store and injected into a component.

Before, they used the `readWriteHOC` that wrapped a component with a stateful component that had all readable and writeable explicitly added to its state. Updates to readable props are achieved by adding a callback via the store's `subscribe` method; this callback updates the state.

Now, in order to reuse the same code paths that implement implicit prop bindings, the `readWriteHOC` was removed. Updates are propagated by the store directly to all components that read data from it; components no longer call `subscribe`.

Parsing explicit props and injecting them into a wrapper around the target component is now done in `withState` and the code branch reuses the `statefulComponentFactory` that handles store retrieval (via `withContext`) and a mock getter.