Add `store.dispatch` and `store.createAction`
This is already enough to offer basic support for actions to implicitly bound components!

Since `store.dispatch` is available to implicitly bound components, they can use it to dispatch actions that they brought into scope or that named. `dispatch` calls actions with a store reference and all additional arguments that were passed to it.

`createAction` takes a function ("proto-action") or a string (name) and a function and returns a function that will immediately dispatch that action. If a name is supplied, the action will be saved internally under that name and can be invoked using `dispatch(name)`.