Change store ref in actions to be readonly
While direct store access is important for the complexity floor of fleux, there are certain situations in which to restrict it. This is the case inside of actions. When the user opts in to the more complex style of using fleux by using actions, he can be expected to respect some additional constraints, if they are in place for good reason. There are at least three reason, why writing store properties from inside an action are a bad idea:

1) It makes the state less predictable. It can easily hide write accesses and thus obscure when and where data is change.
2) It prevents potential future optimizations or at least complicates them. It is likely that future changes to fleux will try to make it smarter in terms of state diffing and triggering rerenders. This is a lot easier when only the return values of actions have to be dealt with.
3) There is no good reason that I can see why a user would want to do this. The user has to write actions such that they return a state slice anyway and while I empathize with users who prefer a more imperative style, I do think that in this case the tradeoff is easily made in favor of a more declarative one.

Write access to store properties will now cause an exception to be thrown. Deep access is not prevented, but that is not considered standard usage anyway.

This is a breaking change.