Change prop renaming order
Before, the order of prop names when renaming props using an object as `withState`'s first argument, was 'store-name: local-name'. This commit reverses this order to be 'local-name: store-name'. This changes is made to be more consistent with the `mapStateToProps` function: Its return value also has the local names as property names. Other than that I do not really have an opinion regarding this order, which is why consistency is the winning argument here.