Add setter injection
The previous refactoring of `parseProps` and how it interacts with `withState` omitted the injection of setter props into components. This commit adds those back into fleux.

But: Since there is no longer a distinction between read and write props, and setters are only injected for explicitly bound props, any change of such a prop will cause the component to rerender even if it only uses the setter, but not the getter prop. I am not sure at this point, whether this can be prevented.

Given that components potentially needlessly rerender if they only access setters, there is not really a point in having setters anyway: To prevent rerenders, the user should rely on actions. I hope, that I can find a solution to offer the old behavior again, without making the API to complicated.

For now, I will add these somewhat less useful setters for better compatibility with older versions of fleux.