Fix implicitly bound components failing on updates
This was an oversight: `selectState` was not properly verified to be a function, which is always is if it went through `parseProps` but which is never is if `withState` was called with no arguments. This was not caught by tests as these were not yet updated to work with the new API anyway.