Add `prepareWriteAction` function to generate setters
The new `prepareWriteAction` function is stand-alone, not a store method. It takes a key name and returns an action function. The action can be dispatched and as an additional argument a setter function should be passed. When the setter gets called, any additional arguments are passed on.

The way these write actions work, they are wrapped by `bindActionProps` such that the result is a setter prop that just receives the actual setter (and additional arguments). `prepareWriteAction` is really meant to be used as part of `bindActionProps` and not on its own (because that's needlessly complicated compared to a simple dispatch or `store.set`).

I am somehow really bad at explaining this function so please refer to the test cases for examples.