Bump version to v0.23.0
The new version v0.23.0 includes the new `selectStateProps` feature, includes breaking changes compared to v0.0.22 and some internal refactoring, and it adds new tests.

I decided to bump to v0.23.0 instead of v0.0.23 to acknowledge the fact that with this change (which was planned for some time) fleux achieves the next level of maturity. The version number should reflect that the most important features are present in this version of fleux and give users confidence that they the library is functional and somewhat stable.