Fix `selectState` not including initial values in store argument
This bug occurred when a `selectState` method read at least one key that was part of the store's initialization and at least one key that was added later. If the key that was added later changed, `selectState` would be evaluated again, but this time without the key(s) from the store initialization.

The reason for this was the way in which keys are created internally in the store: The store's `create` method uses `defineProperty` which defaults to making properties non-enumerable. This caused those keys to not be added to the store ref that was passed to `selectState`.

This commit also adds a test case that would have caught this issue.