Fix `useSelector` causing unnecessary calls to `selectStateProps`
This commit reduces the number of calls to `selectStateProps` that is
happening inside of `useSelector`. This should not cause any necessary
rerenders to take place, just prevent unnecessary calls.

The solution to this surprisingly was to remove most of the callback
that is given to `STORE_GET`. Setting the local state inside of the
callback results in the whole `useSelector` function getting called
again, including its call to `selectStateProps`. I removed the
additional call to `selectStateProps` in the callback, and the
comparison against the previous store state. There no is no state
diffing in my code and I rely on React to compare the return values of
the stateful hook (which does not change if there is no new data).