Fix `unsubscribe` removing callbacks erroneously
Classic JavaScript mistake: When the callback is not found in the list,
`indexOf` returns `-1` which, when passed to `splice` removes the last
element in the array.

In this case, if a callback function was given to `unsubscribe` that
was not subscribed to the given property, it would cause a different
callback (the last one in the list) to be removed instead.