Change bundler from rollup to esbuild
While `esbuild` is still marketed as somewhat experimental, its speed
just does the trick for me. On the laptop I am currently on, the
bundling time went from 3.5s with rollup to 50ms with esbuild! This is
such a significant improvement to the workflow, that I decide in favor
of replacing rollup with esbuild.